backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.8",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "openai>=1.12.0",
    "anthropic>=0.21.0",
    "modelcontextprotocol>=1.0.1",
    "duckduckgo-search>=5.0.0",
    "kafka-python>=2.0.2",
    "dapr>=1.12.0",
    "redis>=5.0.0",
    "apscheduler>=3.10.4",
    "confluent-kafka>=2.3.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
